+++ HandballBackend/Database/Models/Person.cs
@@ -35,7 +35,7 @@ public class Person {
 
     [Required]
     [Column("created_at")]
-    public int? CreatedAt { get; set; } = Utilities.GetUnixSeconds();
+    public long? CreatedAt { get; set; } = Utilities.GetUnixSeconds();
 
     [Required]
     [Column("permission_level")]
